Fix accel error in vel_profiler
The acceleration was always calculated at the maximum step, so on
accel/decel, the code would always be at its final value. This was
sort of noticeable on accel, but it was very noticeable on decel
because the robot was effectively being asked to panic stop, and
it is designed to allow that to happen for safety.